backend/requirements.txt
# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
gunicorn==21.2.0

# MongoDB (motor is async, pymongo for sync)
motor==3.6.0
pymongo==4.9.1

# Auth0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Content extraction (simplified - no lxml)
beautifulsoup4==4.12.2
html5lib==1.1
requests==2.31.0

# Data validation (latest stable with wheels)
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.1.0

# Testing
pytest==7.4.3
httpx==0.26.0
